screen: implement trivial ageing
To fully support the new ageing API, we now implement trivial ageing. On
each modification, we simply increase the screen-age (which overwrites any
line or cell age). This basically invalidates the whole screen on each
change, but at least makes the ageing API work.

Note that for cursor movements we actually age only both changed cells.
Same for flag-changes. But these are the only non-trivial ageing
implementations so far.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>