build: add automated memleak tests
We now use valgrind to perform automated mem-leak tests on selected
test-programs. We also add a separate test_valgrind program which is used
to verify that the memcheck actually works.

We don't run these tests as part of the normal test-suite, however, we
require them for distcheck. Thus, you can avoid using valgrind if you're
not about to do dist-releases. Every developer should have it installed,
anyway, so it's fine.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>