Revert "screen: remove empty cell fallback"
This reverts:
  commit 68e64ecf6d4a0b5d959962d875a97694b6abc23a
  Author: David Herrmann <dh.herrmann@gmail.com>
  Date:   Tue Oct 22 15:53:31 2013 +0200

      screen: remove empty cell fallback

Our assumption was actually not true. If we render scrollback-buffer
lines, we cannot assume that they are long enough. So add the dummy cell
again and use it if a cell is not reachable.

Note that we're safe removing the cursor-fallback, though (which was done
in a follow up). Scroll-back buffers never contain the cursor so we're
fine.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>