gtktsm: add libtsm example
GtkTsmTerminal is a Gtk+-3.0 widget using libtsm. It provides a generic
terminal-widget that can be easily used to create any kinds of
terminal-emulators.

The GtkTsm program is a small example that shows how to use GtkTsm.

Note that all this is optional and only meant as example. Maybe at some
point we will install GtkTsmTerminal as shared object. However, it will
never be mandatory!

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>